cv_detector/new_detector/requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pillow-heif>=0.13.0

# For CNN model
torch>=2.0.0
torchvision>=0.15.0

# For Claude API
anthropic>=0.18.0
